[Feat] Voucher엔티티 VoucherType 필드 추가
