+++ src/main/java/com/prgrms/voucher_manage/domain/voucher/entity/FixedAmountVoucher.java
@@ -1,10 +1,15 @@
 package com.prgrms.voucher_manage.domain.voucher.entity;
 
+import com.prgrms.voucher_manage.console.VoucherType;
+
 import java.util.UUID;
+
+import static com.prgrms.voucher_manage.console.VoucherType.FIXED;
 
 public class FixedAmountVoucher implements Voucher {
     private final UUID voucherId;
     private final Long discountPrice;
+    private final VoucherType voucherType = FIXED;
 
     public FixedAmountVoucher(Long discountPrice){
         this.voucherId = UUID.randomUUID();
@@ -21,4 +26,7 @@ public class FixedAmountVoucher implements Voucher {
     public Long getDiscountAmount() {
         return discountPrice;
     }
+
+    @Override
+    public VoucherType getVoucherType() {return voucherType;}
 }
