+++ src/main/java/com/prgrms/voucher_manage/domain/voucher/entity/PercentAmountVoucher.java
@@ -1,12 +1,17 @@
 package com.prgrms.voucher_manage.domain.voucher.entity;
 
+import com.prgrms.voucher_manage.console.VoucherType;
+
 import java.util.UUID;
+
+import static com.prgrms.voucher_manage.console.VoucherType.*;
 
 public class PercentAmountVoucher implements Voucher {
     private final UUID voucherId;
     private final Long discountPercent;
+    private final VoucherType voucherType = PERCENT;
 
-    public PercentAmountVoucher(Long discountPercent){
+    public PercentAmountVoucher(Long discountPercent) {
         this.voucherId = UUID.randomUUID();
         this.discountPercent = discountPercent;
     }
@@ -20,5 +25,11 @@ public class PercentAmountVoucher implements Voucher {
     public Long getDiscountAmount() {
         return discountPercent;
     }
+
+    @Override
+    public VoucherType getVoucherType() {
+        return voucherType;
+    }
+
 
 }
